requirements.txt
pytest==8.4.1
pytest-asyncio==1.1.0
allure-pytest==2.13.3
allure-python-commons==2.13.3
typing-extensions==4.13.2
mypy==1.15.0
ruff==0.11.10
bandit==1.8.3
uv==0.7.5
prometheus-client==0.21.1